feat(packages): project redesign and rewrite
preset packages are independent; remove preset package in favor of a general web preset;
transpiler/builder are abstracted; add full esm support; esbuild based transpilation/bundling;
update dependencies; drop support for node <16

BREAKING CHANGE: project has been almost entirely rewritten; please check latest documentation

+++ packages/utils/src/tmp-file.ts
@@ -1,5 +1,6 @@
-import fs from 'fs';
 import { Serial, TypeGuard } from 'type-core';
+import fs from 'node:fs';
+
 import { tmpPath } from './tmp-path';
 
 export function tmpFile(
@@ -7,6 +8,15 @@ export function tmpFile(
   content: Serial.Object | Serial.Array | string
 ): string {
   const file = tmpPath(ext, { content });
+
+  let exists = true;
+  try {
+    fs.accessSync(file, fs.constants.F_OK);
+  } catch (_) {
+    exists = false;
+  }
+
+  if (exists) return file;
 
   fs.writeFileSync(
     file,
